+++ Source/WebSites/DevelopmentWebsite/UserControl/HeaderHome.ascx.cs
@@ -130,7 +130,7 @@ public partial class UserControl_HeaderHome : System.Web.UI.UserControl
         {
             topNavigationYM.Visible = true;
             topNavigationYT.Visible = false;
-        }  
+        }
 
         fbWebContext = FacebookWebContext.Current;  // get current facebook session
         StateManager stateManager = StateManager.Instance;
@@ -147,7 +147,7 @@ public partial class UserControl_HeaderHome : System.Web.UI.UserControl
             else if (intUserType == 2)
             {
                 _userName = objSessionvalue.UserName;
-                int NetCreditPoints;
+                double NetCreditPoints;
                 UserRegistration _objUserReg = new UserRegistration();
                 Users objUsers = new Users();
                 objUsers.UserId = objSessionvalue.UserId;
