+++ Source/WebSites/DevelopmentWebsite/UserControl/YourTributeHeader.ascx.cs
@@ -150,7 +150,7 @@ public partial class UserControl_YourTributeHeader : System.Web.UI.UserControl
             else if (intUserType == 2)
             {
                 _userName = objSessionvalue.UserName;
-                int NetCreditPoints;
+                double NetCreditPoints;
                 UserRegistration _objUserReg = new UserRegistration();
                 Users objUsers = new Users();
                 objUsers.UserId = objSessionvalue.UserId;
